:tractor: Move tests that are not related from cobra to parse_test.go
