+++ pkg/cmds/middlewares/middlewares.go
@@ -2,4 +2,6 @@ package middlewares
 
 import "github.com/go-go-golems/glazed/pkg/cmds/layers"
 
-type Handler func(layer *layers.ParameterLayer)
+type HandlerFunc func(layers *layers.ParameterLayers, parsedLayers *layers.ParsedLayers)
+
+type Middleware func(HandlerFunc) HandlerFunc
